requirements.txt
allure-pytest==2.13.5
pytest==8.3.4
selenium==4.29.0
webdriver-manager==4.0.2
faker==37.1.0
pytest-xdist==3.6.1
allure-python-commons~=2.13.5
pydantic==2.11.1
requests==2.32.3
deepdiff==8.4.2